requirements.txt
biopython
pandas
numpy
torch
esm
scikit-learn
joblib
fair-esm